# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Ofertas/GetDetalleOferta": {
      "get": {
        "tags": [
          "Ofertas"
        ],
        "operationId": "GetDetalleOferta",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ofertas/CreateOferta": {
      "post": {
        "tags": [
          "Ofertas"
        ],
        "operationId": "CreateOferta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaForCreationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaForCreationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaForCreationDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reparaciones/GetDetalleReparar": {
      "get": {
        "tags": [
          "Reparaciones"
        ],
        "operationId": "GetDetalleReparar",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reparaciones/CrearReparacion": {
      "post": {
        "tags": [
          "Reparaciones"
        ],
        "operationId": "CrearReparacion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionCreacionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionCreacionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionCreacionDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetalleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "OfertaDetailDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date-time"
          },
          "fechaFinal": {
            "type": "string",
            "format": "date-time"
          },
          "fechaOferta": {
            "type": "string",
            "format": "date-time"
          },
          "metodoPago": {
            "type": "string",
            "nullable": true
          },
          "dirigidaA": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OfertaForCreationDTO": {
        "required": [
          "fechaFinal",
          "fechaInicio",
          "items",
          "metodoPago"
        ],
        "type": "object",
        "properties": {
          "fechaInicio": {
            "type": "string",
            "format": "date-time"
          },
          "fechaFinal": {
            "type": "string",
            "format": "date-time"
          },
          "metodoPago": {
            "minLength": 1,
            "type": "string"
          },
          "dirigidaA": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaItemDTO"
            }
          }
        },
        "additionalProperties": false
      },
      "OfertaItemDTO": {
        "required": [
          "herramientaId",
          "porcentaje"
        ],
        "type": "object",
        "properties": {
          "herramientaId": {
            "type": "integer",
            "format": "int32"
          },
          "porcentaje": {
            "maximum": 100,
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "herramientaNombre": {
            "type": "string",
            "nullable": true
          },
          "herramientaMaterial": {
            "type": "string",
            "nullable": true
          },
          "fabricanteNombre": {
            "type": "string",
            "nullable": true
          },
          "precioOriginal": {
            "type": "number",
            "format": "double"
          },
          "precioFinal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReparacionCreacionDTO": {
        "required": [
          "apellidoCliente",
          "fechaEntrega",
          "itemsReparacion",
          "metodoPago",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "nombreCliente": {
            "maxLength": 20,
            "minLength": 0,
            "pattern": "^[A-Z]+[a-zA-Z''-'\\s]*$",
            "type": "string"
          },
          "apellidoCliente": {
            "maxLength": 40,
            "minLength": 0,
            "pattern": "^[A-Z]+[a-zA-Z''-'\\s]*$",
            "type": "string"
          },
          "numTelefono": {
            "maxLength": 9,
            "minLength": 0,
            "pattern": "^[0-9]*$",
            "type": "string",
            "nullable": true
          },
          "fechaEntrega": {
            "type": "string",
            "format": "date"
          },
          "metodoPago": {
            "$ref": "#/components/schemas/tiposMetodosPago"
          },
          "itemsReparacion": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReparacionItemDTO"
            }
          }
        },
        "additionalProperties": false
      },
      "ReparacionDetalleDTO": {
        "required": [
          "metodoPago"
        ],
        "type": "object",
        "properties": {
          "metodoPago": {
            "$ref": "#/components/schemas/tiposMetodosPago"
          },
          "reparacionId": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellidoCliente": {
            "type": "string",
            "nullable": true
          },
          "fechaEntrega": {
            "type": "string",
            "format": "date-time"
          },
          "fechaRecogida": {
            "type": "string",
            "format": "date-time"
          },
          "numTelefono": {
            "type": "string",
            "nullable": true
          },
          "itemsReparacion": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReparacionItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReparacionItemDTO": {
        "type": "object",
        "properties": {
          "herramientaID": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "precioReparacion": {
            "type": "number",
            "format": "double"
          },
          "tiempoReparacion": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "tiposMetodosPago": {
        "enum": [
          "TarjetaCredito",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      }
    }
  }
}